Added `render_as_GET_url()`, `sandbox_as_GET_url()` methods and 'image_url' header image support to form.
The method `<PayPalPaymentForm>.render_as_GET_url()` and it's sandbox version can be used
to directly redirect to PayPal payment page from view that performs pre-payment processing;
for example a view that check if shipping information was entered and is correct, to confirming
that inventory is still adequate for order. Readme documentation was updated to reflect these changes.